rework import style
