--- libs/design/card/examples/src/examples.ts
@@ -1,11 +1,0 @@
-import { BasicCardComponent } from './basic-card/basic-card.component';
-import { CardThemingComponent } from './card-theming/card-theming.component';
-import { LinkableCardComponent } from './linkable-card/linkable-card.component';
-import { RaisedCardComponent } from './raised-card/raised-card.component';
-
-export const CARD_EXAMPLES = [
-  BasicCardComponent,
-  CardThemingComponent,
-  RaisedCardComponent,
-  LinkableCardComponent,
-];
