+++ libs/design/card/examples/src/public_api.ts
@@ -1,8 +1,15 @@
-export { BasicCardComponent } from './basic-card/basic-card.component';
-export { CardThemingComponent } from './card-theming/card-theming.component';
-export { RaisedCardComponent } from './raised-card/raised-card.component';
-export { LinkableCardComponent } from './linkable-card/linkable-card.component';
+import { BasicCardComponent } from './basic-card/basic-card.component';
+export { BasicCardModule } from './basic-card/basic-card.module';
+import { CardThemingComponent } from './card-theming/card-theming.component';
+export { CardThemingModule } from './card-theming/card-theming.module';
+import { LinkableCardComponent } from './linkable-card/linkable-card.component';
+export { LinkableCardModule } from './linkable-card/linkable-card.module';
+import { RaisedCardComponent } from './raised-card/raised-card.component';
+export { RaisedCardModule } from './raised-card/raised-card.module';
 
-export { CardExamplesModule } from './card-examples.module';
-
-export { CARD_EXAMPLES } from './examples';
+export const CARD_EXAMPLES = [
+  BasicCardComponent,
+  CardThemingComponent,
+  LinkableCardComponent,
+  RaisedCardComponent,
+];
